+++ ROS2ESPBridge/ir_driver.h
@@ -3,7 +3,7 @@
   #define IRRIGHT D14
 #endif
 
-int getIRDist(int ir_pin);
+double getIRDist(int ir_pin);
 
 void IRAM_ATTR IR_R_ISR();
 void IRAM_ATTR IR_L_ISR();
